pyproject.toml
[project]
name = "Devious-WinRM"
version = "0.1.0"
description = "A Pentester's (work-in-progress) Powershell Client."
readme = "README.md"
authors = [
    {name = "Pablo Comino", email = "pablo@pablocomino.com"}
]
license = {text = "GPL-3.0-or-later"}
dependencies = [
    "prompt_toolkit==3.0.5",
    "pypsrp[kerberos]>=1.0.0b1,<1.1.0",
    "impacket==0.12.0",
    "Pygments==2.19.1"
]

[project.scripts]
devious-winrm = "devious_winrm.cli:cli"
dwrm = "devious_winrm.cli:cli"

[project.entry-points."pipx.run"]
devious_winrm = "devious_winrm.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
select = ["ALL"]
pyupgrade.keep-runtime-typing = true
ignore = ["RUF013", "PERF203"]
lint.pylint.allow-magic-value-types = ["int", "str"]
